Employee Performance Mapping
To facilitate a better understanding, managers have provided ratings for each
employee which will help the HR department to finalize the employee performance
mapping. As a DBA, you should find the maximum salary of the employees and
ensure that all jobs are meeting the organization's profile standard. You also need
to calculate bonuses to find extra cost for expenses. This will raise the overall
performance of the organization by ensuring that all required employees receive
training.